+++ src/components/Widget/GoogleReCaptchaWidget.jsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import { useIntl } from 'react-intl';
+import { Grid } from 'semantic-ui-react';
+import { injectLazyLibs } from '@plone/volto/helpers/Loadable/Loadable';
+
+const GoogleReCaptchaWidget = ({ onVerify, GoogleReCaptcha }) => {
+  const intl = useIntl();
+  const {
+    GoogleReCaptcha: ReCaptcha,
+    GoogleReCaptchaProvider,
+  } = GoogleReCaptcha;
+
+  return (
+    <GoogleReCaptchaProvider
+      reCaptchaKey={process.env.RAZZLE_RECAPTCHA_KEY}
+      language={intl.locale ?? 'en'}
+    >
+      <Grid.Row centered className="row-padded-top">
+        <Grid.Column textAlign="center">
+          <ReCaptcha onVerify={onVerify} />
+        </Grid.Column>
+      </Grid.Row>
+    </GoogleReCaptchaProvider>
+  );
+};
+
+export default injectLazyLibs(['GoogleReCaptcha'])(GoogleReCaptchaWidget);
